requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# UI
streamlit>=1.29.0

# Utilities (standard library alternatives exist but these are cleaner)
python-dateutil>=2.8.0

# Development/testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
